+++ plugin-packs/postcss-bundler/.tape.mjs
@@ -5,6 +5,10 @@ postcssTape(plugin)({
 	basic: {
 		message: "supports basic usage",
 	},
+	'leading-slash': {
+		message: "does not infer a root to resolve leading slash imports",
+		exception: /Failed to find \'\/imports\/basic.css\'/,
+	},
 	'examples/example': {
 		message: 'minimal example',
 	},