+++ packages/postcss-tape/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changes to PostCSS Tape
 
+### Unreleased (patch)
+
+- Fix processor options
+
 ### 3.0.0
 
 _July 3, 2023_